Add builder pattern for RDS
