+++ pkg/envoy/rds/rbac_test.go
@@ -1,4 +1,4 @@
-package route
+package rds
 
 import (
 	"fmt"
