+++ pkg/envoy/rds/response.go
@@ -6,7 +6,6 @@ import (
 	"github.com/openservicemesh/osm/pkg/catalog"
 	"github.com/openservicemesh/osm/pkg/certificate"
 	"github.com/openservicemesh/osm/pkg/envoy"
-	"github.com/openservicemesh/osm/pkg/envoy/rds/route"
 	"github.com/openservicemesh/osm/pkg/envoy/registry"
 	"github.com/openservicemesh/osm/pkg/errcode"
 	"github.com/openservicemesh/osm/pkg/trafficpolicy"
@@ -17,10 +16,9 @@ func NewResponse(cataloger catalog.MeshCataloger, proxy *envoy.Proxy, cm *certif
 	proxyServices, err := cataloger.ListServicesForProxy(proxy)
 	if err != nil {
 		log.Error().Err(err).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrFetchingServiceList)).
-			Msgf("Error looking up services for proxy %s", proxy)
+			Msgf("Error looking up services for Envoy with name=%s", proxy)
 		return nil, err
 	}
-	var rdsResources []types.Resource
 
 	trustDomain := cm.GetTrustDomain()
 
@@ -32,35 +30,24 @@ func NewResponse(cataloger catalog.MeshCataloger, proxy *envoy.Proxy, cm *certif
 		}
 	}
 
-	// ---
-	// Build inbound mesh route configurations. These route configurations allow
-	// the services associated with this proxy to accept traffic from downstream
-	// clients on allowed routes.
+	routesBuilder := RoutesBuilder().
+		Proxy(proxy).
+		StatsHeaders(statsHeaders).
+		TrustDomain(trustDomain)
+
+	// Get inbound mesh traffic policy and pass to builder
 	inboundMeshTrafficPolicy := cataloger.GetInboundMeshTrafficPolicy(proxy.Identity, proxyServices)
 	if inboundMeshTrafficPolicy != nil {
-		inboundMeshRouteConfig := route.BuildInboundMeshRouteConfiguration(inboundMeshTrafficPolicy.HTTPRouteConfigsPerPort, proxy, statsHeaders, trustDomain)
-		for _, config := range inboundMeshRouteConfig {
-			rdsResources = append(rdsResources, config)
-		}
+		routesBuilder.InboundPortSpecificRouteConfigs(inboundMeshTrafficPolicy.HTTPRouteConfigsPerPort)
 	}
 
-	// ---
-	// Build outbound mesh route configurations. These route configurations allow this proxy
-	// to direct traffic to upstream services that it is authorized to connect to on allowed
-	// routes.
+	// Get outbound mesh traffic policy and pass to builder
 	outboundMeshTrafficPolicy := cataloger.GetOutboundMeshTrafficPolicy(proxy.Identity)
-
 	if outboundMeshTrafficPolicy != nil {
-		outboundMeshRouteConfig := route.BuildOutboundMeshRouteConfiguration(outboundMeshTrafficPolicy.HTTPRouteConfigsPerPort)
-		for _, config := range outboundMeshRouteConfig {
-			rdsResources = append(rdsResources, config)
-		}
+		routesBuilder.OutboundPortSpecificRouteConfigs(outboundMeshTrafficPolicy.HTTPRouteConfigsPerPort)
 	}
 
-	// ---
-	// Build ingress route configurations. These route configurations allow the
-	// services associated with this proxy to accept ingress traffic from downstream
-	// clients on allowed routes.
+	// Get ingress traffic policies and pass to builder
 	var ingressTrafficPolicies []*trafficpolicy.InboundTrafficPolicy
 	for _, svc := range proxyServices {
 		ingressPolicy, err := cataloger.GetIngressTrafficPolicy(svc)
@@ -74,23 +61,23 @@ func NewResponse(cataloger catalog.MeshCataloger, proxy *envoy.Proxy, cm *certif
 		}
 		ingressTrafficPolicies = trafficpolicy.MergeInboundPolicies(ingressTrafficPolicies, ingressPolicy.HTTPRoutePolicies...)
 	}
-	if len(ingressTrafficPolicies) > 0 {
-		ingressRouteConfig := route.BuildIngressConfiguration(ingressTrafficPolicies, trustDomain)
-		rdsResources = append(rdsResources, ingressRouteConfig)
-	}
+	routesBuilder.IngressTrafficPolicies(ingressTrafficPolicies)
 
-	// ---
-	// Build egress route configurations. These route configurations allow this
-	// proxy to direct traffic to external non-mesh destinations on allowed routes.
+	// Get egress traffic policy and pass to builder
 	egressTrafficPolicy, err := cataloger.GetEgressTrafficPolicy(proxy.Identity)
+
 	if err != nil {
 		log.Error().Err(err).Msgf("Error retrieving egress traffic policies for proxy with identity %s, skipping egress route configuration", proxy.Identity)
 	}
 	if egressTrafficPolicy != nil {
-		egressRouteConfigs := route.BuildEgressRouteConfiguration(egressTrafficPolicy.HTTPRouteConfigsPerPort)
-		for _, egressConfig := range egressRouteConfigs {
-			rdsResources = append(rdsResources, egressConfig)
-		}
+		routesBuilder.EgressPortSpecificRouteConfigs(egressTrafficPolicy.HTTPRouteConfigsPerPort)
+	}
+
+	rdsResources, err := routesBuilder.Build()
+
+	if err != nil {
+		log.Error().Err(err).Msgf("Error building route config.")
+		return nil, err
 	}
 
 	return rdsResources, nil
