--- pkg/envoy/rds/route/suite_test.go
@@ -1,13 +1,0 @@
-package route
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestRouteConfiguration(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Route Configuration Test Suite")
-}
