--- pkg/envoy/rds/route/types.go
@@ -1,10 +1,0 @@
-// Package route implements utility routines to build HTTP route configurations for Envoy proxies.
-package route
-
-import (
-	"github.com/openservicemesh/osm/pkg/logger"
-)
-
-var (
-	log = logger.New("envoy/route")
-)
